requirements.txt
Appium-Python-Client==5.2.4
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
outcome==1.3.0.post0
pillow==12.0.0
PySocks==1.7.1
python-dotenv==1.2.1
selenium==4.39.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.0
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2